resstructure the books controller
