+++ backend/api/controllers/books/index.js
@@ -1,0 +1,19 @@
+const addBook = require('./addBook');
+const deleteBook = require('./deleteBook');
+const getAllBooksByUserId = require('./getAllBooksByUserId');
+const getAllBooks = require('./getAllBooks');
+const getBookById = require('./getBookById');
+const getBookDetails = require('./getBookDetails');
+const searchBooks = require('./searchBooks');
+const updateBook = require('./updateBook');
+
+module.exports = {
+  addBook,
+  deleteBook,
+  getAllBooksByUserId,
+  getAllBooks,
+  getBookById,
+  getBookDetails,
+  searchBooks,
+  updateBook,
+};
